feat: calculate for Premium & Verify email

+++ Domain/src/Control/Sale/Sale.java
@@ -137,7 +137,8 @@ public class Sale {
 		contract.setCustomerId(customer.getId());
 		contract.setInsuranceId(insuranceId);
 		contract.setPremiumRate(this.insuranceList.get(insuranceId).getPremium());
-		contract.setInsurancePrice(this.calculatePremium.calculatePremium(customer, this.insuranceList.get(insuranceId).getPremium()));
+		contract.setInsurancePrice(
+				(int) this.calculatePremium.calculatePremium(customer, this.insuranceList.get(insuranceId).getPremium()));
 
 		this.contractList.add(contract);
 	}
